Add ks_util
